feat(documents): Add document validation logic and helper functions for file uploads
